corregido errores en MatchOne

+++ examples/MatchOne/src/main/java/com/ilargia/games/components/Input.java
@@ -5,10 +5,10 @@ import com.ilargia.games.entitas.codeGenerator.Component;
 
 @Component(pools = {"Input"})
 public class Input implements IComponent {
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
-    public Input(float x, float y) {
+    public Input(int x, int y) {
         this.x = x;
         this.y = y;
     }
